Doc: Usuarios y Ordenes Funcionando

+++ backend/controllers/productoController.js
@@ -3,7 +3,7 @@ const Producto = require('../models/productoModel')
 
 const getProductos = asyncHandler( async (req, res) => {
 
-    const productos = await Producto.find()
+    const productos = await Producto.find({user: req.user.id})
 
     // res.status(200).json({mensaje: 'Mostrar los productos'})
     res.status(200).json(productos)
@@ -11,21 +11,21 @@ const getProductos = asyncHandler( async (req, res) => {
 
 const setProductos = asyncHandler( async (req, res) => {
     // console.log(req.body)
-    if(!req.body.nombre || (!req.body.categoria) || (!req.body.marca) || (!req.body.sku) || (!req.body.precio)){
+    if(!req.body.nombre || (!req.body.categoria) || (!req.body.marca) || (!req.body.precio)){
         // res.status(400).json({mensaje: 'Favor de teclear la descripcion del producto'})
         res.status(400)
         throw new Error('Favor de teclear la descripcion del producto ')
     }
     // res.status(201).json({mensaje: 'Crear un Producto'})
-    const tarea = await Producto.create({
+    const producto = await Producto.create({
         nombre: req.body.nombre,
         categoria: req.body.categoria,
         marca: req.body.marca,
-        sku: req.body.sku,
         precio: req.body.precio,
+        user: req.user.id
     })
 
-    res.status(201).json(tarea)
+    res.status(201).json(producto)
 })
 
 const updateProductos = asyncHandler( async (req, res) => {
@@ -35,6 +35,12 @@ const updateProductos = asyncHandler( async (req, res) => {
         res.status(400)
         throw new Error('Producto no encontrado')
     }
+
+    if(producto.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso No Autorizado, el prodcto no pertenece al usuario logeado')
+    }
+
     const productoModificado = await Producto.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json(productoModificado)
@@ -48,6 +54,12 @@ const deleteProductos = asyncHandler( async (req, res) => {
         res.status(400)
         throw new Error('Producto no encontrado')
     }
+
+    if(producto.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso No Autorizado, el producto no pertenece al usuario logeado')
+    }
+
     await producto.deleteOne()
 
     res.status(200).json({id: req.params.id})
